Add demo data
