+++ scalemae/demo/__init__.py
@@ -1,0 +1,8 @@
+"""
+mkinit ~/code/watch/geowatch_tpl/submodules/scale-mae/scalemae/demo/__init__.py
+"""
+from scalemae.demo import kwcoco_demodata
+
+from scalemae.demo.kwcoco_demodata import (make_demo_image_folder,)
+
+__all__ = ['kwcoco_demodata', 'make_demo_image_folder']
