+++ scalemae/util/datasets.py
@@ -16,46 +16,15 @@ from timm.data.constants import IMAGENET_DEFAULT_MEAN, IMAGENET_DEFAULT_STD
 from torchvision import datasets, transforms
 
 
-def make_demo_image_folder():
-    import kwcoco
-    import kwimage
-    import ubelt as ub
-    dpath = ub.Path.appdir('scalemae/tests/demo/imagefolder')
-    stamp = ub.CacheStamp(fname='demo-images', dpath=dpath)
-    if stamp.expired():
-        dset = kwcoco.CocoDataset.demo('vidshapes8')
-        for coco_img in ub.ProgIter(dset.images().coco_images, desc='setup images'):
-            image_id = coco_img['id']
-
-            img = coco_img.imdelay()
-
-            anns = dset.annots(image_id=image_id).objs
-            for ann in anns:
-                # Crop out the annotation
-                annot_id = ann['id']
-                box = kwimage.Box.coerce(ann['bbox'], format='xywh')
-                sl = box.quantize().to_slice()
-                crop = img[sl]
-                resized = crop.resize((224, 224))
-                crop_imdata = resized.finalize()
-
-                # Save in a folder with its category name
-                category_name = dset.index.cats[ann['category_id']]['name']
-                cat_dpath = (dpath / category_name).ensuredir()
-                crop_fpath = cat_dpath / f'annot_{annot_id:03d}.jpg'
-                kwimage.imwrite(crop_fpath, crop_imdata)
-        stamp.renew()
-    return dpath
-
-
 def build_dataset(is_train, args, explicit_path=False):
     """
     Example:
         >>> # xdoctest: +REQUIRES(module:kwcoco)
         >>> # Hack: use a scriptconfig object to simulate args
         >>> from scalemae.util.datasets import *  # NOQA
+        >>> from scalemae import demo
         >>> import scriptconfig as scfg
-        >>> data_path = make_demo_image_folder()
+        >>> data_path = demo.make_demo_image_folder()
         >>> class DatasetArgs(scfg.DataConfig):
         >>>     data_path = None
         >>>     input_size = 224
@@ -67,10 +36,20 @@ def build_dataset(is_train, args, explicit_path=False):
         >>> args = DatasetArgs(data_path=data_path)
         >>> is_train = True
         >>> dataset = build_dataset(is_train, args, explicit_path=True)
-        >>> for idx in range(len(dataset)):
-        >>>     batch_item = dataset[idx]
+        >>> item1 = dataset[0]
+        >>> itemN = dataset[len(dataset) - 1]
+        >>> # xdoctest: +REQUIRES(--show)
+        >>> import kwplot
+        >>> import kwimage
+        >>> kwplot.autompl()
+        >>> hwc0 = item1[0].permute(1, 2, 0).numpy()
+        >>> hwcN = itemN[0].permute(1, 2, 0).numpy()
+        >>> kwplot.imshow(kwarray.normalize(hwc0), pnum=(2, 2, 1), fnum=1)
+        >>> kwplot.imshow(kwarray.robust_normalize(hwc0), pnum=(2, 2, 2), fnum=1)
+        >>> kwplot.imshow(kwarray.normalize(hwcN), pnum=(2, 2, 3), fnum=1)
+        >>> kwplot.imshow(kwarray.robust_normalize(hwcN), pnum=(2, 2, 4), fnum=1)
+        >>> kwplot.show_if_requested()
     """
-
     # ---
     # JPC: This has a problem, there should have been a config passed here.
     # Not sure where it went but going to fudge it in.
